This must be an int, not a float.

+++ syncthing_gtk/identicon.py
@@ -64,7 +64,7 @@ class IdentIcon(Gtk.DrawingArea):
 		rect_size	= min(allocation.width, allocation.height) / self.size
 		offset_x	= (allocation.width / 2) - (rect_size * self.size / 2)
 		offset_y	= (allocation.height / 2) - (rect_size * self.size / 2)
-		middle_col	= self.size / 2
+		middle_col	= int(self.size / 2)
 		
 		# Set color
 		cr.set_source_rgba(*self.color)
